------------------------------------------------------------------------ r2920 | horda03 | 2010-04-30 16:30:19 +0100 (Fri, 30 Apr 2010) | 24 lines
Where Ingresdate Intervals and ANSI intervals (year to month, day to second)
shared "format" characters in the conversion of a string to the interval,
the value for the Ingresdate was being negated when it shouldn't.

For Ingresdate intervals the values should be passed through as it

   '-10 hrs 22 mins' should be stored as '-9 hrs -38 mins' (all parts
negative), but in reality the value '-10 hrs -22 mins' was stored.

Now negative_interval will only be defined to TRUE for ANSI intervals
where the Year format is 'x' or the Day format is 'e'. Only when this
flag is set (or the format character indicates negation required)
will the value assigned be multiplied by -1.

In investigating bug 123654, I stumbled upon another related problem
(from hence force known as bug 123666).

The result of -interval_dtos('-38765 23:59:59.0') was
-2778441 00:00:01.000000000, not quite what i expected.
It turned out that the negative interval code was using the DATE
normalisation routine. Corrected to use the Interval normalisation function.

------------------------------------------------------------------------
r2921 | jenjo02 | 2010-04-30 18:15:28 +0100 (Fri, 30 Apr 2010) | 17 lines

SD 142262

Solaris segmentation violation in dm0pUnpinBuf when running the CTS tests,
when a page deadlock occurred in dm1b_allocate().

Fortunately reproduced on ingdevserver1.

Caused by incomplete initialization of local DMP_PINFO structures
in dm2r_replace().

Though pinfo.page was NULL, garbage in pinfo.pincount induced a call
to unpin the NULL page, causing the SEGV.

One must ensure that all DMP_PINFO structures are fully initialized
by way of the DMP_PINIT macro.

------------------------------------------------------------------------