------------------------------------------------------------------------ r2934 | bonro01 | 2010-05-03 23:00:09 +0100 (Mon, 03 May 2010) | 4 lines
Fix Windows link problem caused by a previous change.
Remove references because of change 504290 for bug 121123

------------------------------------------------------------------------
r2935 | kiria01 | 2010-05-04 09:46:13 +0100 (Tue, 04 May 2010) | 8 lines

Prior to scalar subquery support the APP node of the query plan was always
at the head.  In the case in question this is not the case thus attaching
the sequence descriptor needs to be done via .ops_cstate.opc_subqry
instead of .ops_osubquery.

Also removed the unhelpful #if 0'd out blocks of code that served no purpose
at all.

------------------------------------------------------------------------