------------------------------------------------------------------------ r2957 | schka01 | 2010-05-07 21:15:11 +0100 (Fri, 07 May 2010) | 7 lines
Encrypted columns were not being handled properly by
standard data compression.  An encrypted column should
never be considered for compression, and if nullable
is not a candidate for null-compression.  (The null
indicator is encrypted along with the rest of the value.)

------------------------------------------------------------------------
r2958 | schka01 | 2010-05-07 21:30:09 +0100 (Fri, 07 May 2010) | 7 lines

The total row length (cp_row_length) was too large because
it was being incremented per column instead of per group.
Unix / windows doesn't use cp_row_length at all, but on
VMS it's used to set the RMS record size for fixed width
copies (such as copy table() into).  Fix the length calculation.

------------------------------------------------------------------------