------------------------------------------------------------------------ r2960 | jenjo02 | 2010-05-07 22:45:10 +0100 (Fri, 07 May 2010) | 18 lines
The fix for bug 123649 to reenable group commit causes a ~50% performance
hit when optimized log writes are enabled and there is a single logwriter
thread defined (the optimum configuration).

Debugging sessions revealed that the most observed long wait for group
commit to do its thing occurred during Btree split when writing a
forced DM0LEMINI log record; while waiting for the group commit
thread to schedule and write this log record, the waiting transaction
holds an exclusive lock on the ROOT page of the index, blocking other
transactions that need access to the Btree.

This change alters LGwrite and LGforce to not induce group
commit when optimized writes are enabled and the number
of logwriters in the installation is one.

With this change in place, performance was restored.

------------------------------------------------------------------------
r2961 | vande02 | 2010-05-08 00:00:10 +0100 (Sat, 08 May 2010) | 4 lines

Branching new SED mask file /be/util/data/utilmask related to MAIN
Change 504454 in ingres on 07-may-2010.

------------------------------------------------------------------------