------------------------------------------------------------------------ r2965 | hanal04 | 2010-05-11 09:45:10 +0100 (Tue, 11 May 2010) | 10 lines
Cross integration of change 504503

The creation of IIDBDB does not establish the dbservice value
in the CNF file. INFODB therefore reports incorrect details.
This change calls iiqef_alter_db() to flush the known dbservice
value into the CNF file just before we exit createdb. The
change is only activated if we are creating IIDBDB.

------------------------------------------------------------------------