------------------------------------------------------------------------ r2967 | schka01 | 2010-05-11 12:30:45 +0100 (Tue, 11 May 2010) | 12 lines
Fix an off-by-one error in MHround() that rounded to
zero one position too soon.
Add FI's and handler functions for floating floor,
ceil, and trunc.  Previously, the decimal versions
of these functions were being used for float, leading
to inaccurate results in some cases.
(I was unable to provoke an incorrect trunc() result
using the old coerce-to-decimal way, but I added
the floating trunc anyway;  it's more likely to be
correct for all inputs.)

------------------------------------------------------------------------