------------------------------------------------------------------------ r2969 | drivi01 | 2010-05-11 17:15:10 +0100 (Tue, 11 May 2010) | 11 lines
Due to recent changes in the way dayfile is built, the DBA Tools
image creation is broken b/c dayfile.dst is missing.

This change updates the DBA Tools installer to
include dayfile only. dayfile.dst is being removed from
distribution.

This is done in a separate change from 32-bit and 64-bit
installers.

------------------------------------------------------------------------