------------------------------------------------------------------------ r2972 | schka01 | 2010-05-12 11:15:09 +0100 (Wed, 12 May 2010) | 8 lines
The previous submission for 123712 used the trunc() C library function,
which apparently does not exist on Solaris 8 or on Windows.
Use cast to integer and back to float8 instead, being careful to
avoid float values larger than an i8 can hold.  (Since an i8 has
more integer bits than float does, such large floats can't have
any decimal places in it anyway.)

------------------------------------------------------------------------
r2973 | coomi01 | 2010-05-12 14:00:14 +0100 (Wed, 12 May 2010) | 27 lines

NMlocksyms() requests SIcreate() to create a lock file.
This may fail due to that file already existing, or
because the user does not have permissions to perform
the operation.

In the case of the latter, the loop waiting for some
other process to give up the lock is in appropriate.

There is a bit of an issue with status codes returned
by SIcreate(), the values of which are defined in
private header silocal.h, and therefore can not be
used from outside the SI sub-utility.

The solution chosen is to take these #defines out
of silocal.h and put them in the main si.h.

We then refine SI_CANT_OPEN with SI_CANT_OPEN_EACESS
and SI_CANT_OPEN_EEXIST. The first meaning unable to
open the file due to lack of priviledge. The second
we can't open file exclusive for it already exists.

The error message now produced by return of NM_STOPN is

  E_CL1405_NM_STOPN       NM[sg]tIngAt: Unable to open symbol table
  only the Ingres super-user can use ingsetenv

------------------------------------------------------------------------