------------------------------------------------------------------------ r2973 | coomi01 | 2010-05-12 14:00:14 +0100 (Wed, 12 May 2010) | 27 lines
NMlocksyms() requests SIcreate() to create a lock file.
This may fail due to that file already existing, or
because the user does not have permissions to perform
the operation.

In the case of the latter, the loop waiting for some
other process to give up the lock is in appropriate.

There is a bit of an issue with status codes returned
by SIcreate(), the values of which are defined in
private header silocal.h, and therefore can not be
used from outside the SI sub-utility.

The solution chosen is to take these #defines out
of silocal.h and put them in the main si.h.

We then refine SI_CANT_OPEN with SI_CANT_OPEN_EACESS
and SI_CANT_OPEN_EEXIST. The first meaning unable to
open the file due to lack of priviledge. The second
we can't open file exclusive for it already exists.

The error message now produced by return of NM_STOPN is

  E_CL1405_NM_STOPN       NM[sg]tIngAt: Unable to open symbol table
  only the Ingres super-user can use ingsetenv

------------------------------------------------------------------------
r2974 | coomi01 | 2010-05-12 14:00:26 +0100 (Wed, 12 May 2010) | 7 lines

When skipping a PST_SEQOP element whilst building
the attributes array, we need to decrement the number
of attributes present. Failure to do so may lead to
SegV's latter on should there be an attempt to read
a non existant entry.

------------------------------------------------------------------------