------------------------------------------------------------------------ r2974 | coomi01 | 2010-05-12 14:00:26 +0100 (Wed, 12 May 2010) | 7 lines
When skipping a PST_SEQOP element whilst building
the attributes array, we need to decrement the number
of attributes present. Failure to do so may lead to
SegV's latter on should there be an attempt to read
a non existant entry.

------------------------------------------------------------------------
r2975 | stial01 | 2010-05-12 15:00:31 +0100 (Wed, 12 May 2010) | 13 lines

SIR 121619 MVCC

E_DM966F_DMVE_BTOVFL_NOROOM was reported during mvcc undo.
Diagnostics in II_DBMS_LOG showed that there was room on the leaf. There was
one DELETED key that could have been cleaned except that the current cursor
was positioned on it.

1) Changed dmve_bid_check to allow any key to be 'cleaned' if the key the
current cursor is positioned on is the same as the key we are putting back.
2) Use DMP_PINIT to fully init local DMP_PINFO structures
3) Minor changes to TRdisplay for reposition failures in dm1b/dmve

------------------------------------------------------------------------