------------------------------------------------------------------------ r2975 | stial01 | 2010-05-12 15:00:31 +0100 (Wed, 12 May 2010) | 13 lines
SIR 121619 MVCC

E_DM966F_DMVE_BTOVFL_NOROOM was reported during mvcc undo.
Diagnostics in II_DBMS_LOG showed that there was room on the leaf. There was
one DELETED key that could have been cleaned except that the current cursor
was positioned on it.

1) Changed dmve_bid_check to allow any key to be 'cleaned' if the key the
current cursor is positioned on is the same as the key we are putting back.
2) Use DMP_PINIT to fully init local DMP_PINFO structures
3) Minor changes to TRdisplay for reposition failures in dm1b/dmve

------------------------------------------------------------------------
r2976 | bonro01 | 2010-05-12 21:45:12 +0100 (Wed, 12 May 2010) | 3 lines

Build sep.1 for Unix and Windows

------------------------------------------------------------------------