------------------------------------------------------------------------ r3000 | vande02 | 2010-05-13 01:30:08 +0100 (Thu, 13 May 2010) | 3 lines
Cross-integrating MAIN Change 503524 to OPENSRCQA.

------------------------------------------------------------------------
r3001 | schka01 | 2010-05-13 14:45:13 +0100 (Thu, 13 May 2010) | 7 lines

The query analyzer for hash join/agg space reservation assumed
that the QEA_NODEACT flag meant that there was really a query
plan under the action.  The flag was being set even if the
"query plan" was null.  Arrange to only set the flag if there
is a non-empty query plan under the action.

------------------------------------------------------------------------