------------------------------------------------------------------------ r3010 | kiria01 | 2010-05-14 11:00:12 +0100 (Fri, 14 May 2010) | 7 lines
REGR_COUNT second parameter mis-compiled.

Derive the second parameter type from the right subtree and
not the left. (first param)

------------------------------------------------------------------------
r3011 | hanal04 | 2010-05-14 15:45:09 +0100 (Fri, 14 May 2010) | 17 lines

Cross integration of change 504594

E_US0010 when tracked using exception handling leads to multiple
inglogs instances running at the same time. This leads to a
conflict on the LOGDIR and the scripts hang. As they do not
return the server is also hung.
This change adds _pid<pid> to the end of the value of LOGDIR
on unix and adds milliseconds to the time portion of the DNAME
in the Windows batch file ($$ equivalent not available in Windows
batch files).

Cross integration of change 504595

Bump version numbers for inglogs as requested by Kristoff.

------------------------------------------------------------------------