------------------------------------------------------------------------ r3013 | maspa05 | 2010-05-14 17:00:12 +0100 (Fri, 14 May 2010) | 16 lines
Later versions of flex on Windows need the "--nounistd" flag so that they
don't try to include the non-existent unistd.h file. A previous change to
Jamdefs.int_w32 catered for this but assumed that Cygwin would always have
the later version.

This change uses an environment variable, FLEX_FLAG, to hold the flag if it
is needed.

bldenv.bat has been amended to set FLEX_FLAG based on the detected version of
flex, unless FLEX_FLAG has been set already by the user.

The default value - if not set by bldenv.bat or elsewhere - is --nounistd
which works with flex 2.5.31 and later. For earlier versions (typically
2.5.4) where bldenv.bat is not run, the user should set FLEX_FLAG={space}

------------------------------------------------------------------------
r3014 | stephenb | 2010-05-17 12:00:10 +0100 (Mon, 17 May 2010) | 4 lines

Fix typo in previous CS change which caused build failures on
non GCC 4.1 platforms.

------------------------------------------------------------------------