------------------------------------------------------------------------ r3016 | kiria01 | 2010-05-17 17:06:47 +0100 (Mon, 17 May 2010) | 6 lines
When injecting a SINGLETON aggregate, only mark the aggregate operator nodes
with a join ID if there are correlated variables that will cause the join ID
to be represented in the inner and outer ranve variable bitmaps. Otherwise,
OPF will see this node as inconsistant due to it not having a real correlation.

------------------------------------------------------------------------
r3017 | loera01 | 2010-05-17 21:30:13 +0100 (Mon, 17 May 2010) | 16 lines

Cross integration of change 504652 from ingres2006r3.

The CvtCharFloat() function, like many other numeric to char conversion
functions, calls CvtCharNumStr() to verify that the string is numeric and to
edit leading and trailing blanks.  Ultimately, CvtCharFloat() invokes CVaf()
to convert the numeric string to a float, but CVaf() requires a null-terminated
string, because there is no length argument.

I decided to make the fix in CvtCharNumStr() because several other ODBC
conversion routines invoke CvtCharNumStr() and would also require a properly
terminated numeric string.  I copied the input string into a work area and set
the null terminator by the string length argument.  This will also truncate
null-terminated strings with length indicators less than the string length,
which is also a desirable behavior.

------------------------------------------------------------------------