------------------------------------------------------------------------ r3025 | kiria01 | 2010-05-19 12:20:59 +0100 (Wed, 19 May 2010) | 16 lines
Addressing bug 123442:
Calls to psl_mk_const_similar() that directly handle default values are
changed to force that routine to build and return a parse tree with the appropriate
coercion operator applied to make the constant match the attribute being defaulted.
This ensures that the types remain consistant when otherwise a DB_LTXT_TYPE would be
left in the parse tree.

This additional behaviour of psl_mk_const_similar() is triggered if the 'handled'
result parameter is passed as NULL. This is interpreted as handle it anyway and the
parse tree is built and the constant folding tried as before but now if the collapse
to constant didn't happen we pass the tree through anyway as the result will be the
same.

In addition, added the missing DB_LTXT_TYPE in the BYTE(ALL) function.

------------------------------------------------------------------------