------------------------------------------------------------------------ r3025 | kiria01 | 2010-05-19 12:20:59 +0100 (Wed, 19 May 2010) | 16 lines
Addressing bug 123442:
Calls to psl_mk_const_similar() that directly handle default values are
changed to force that routine to build and return a parse tree with the appropriate
coercion operator applied to make the constant match the attribute being defaulted.
This ensures that the types remain consistant when otherwise a DB_LTXT_TYPE would be
left in the parse tree.

This additional behaviour of psl_mk_const_similar() is triggered if the 'handled'
result parameter is passed as NULL. This is interpreted as handle it anyway and the
parse tree is built and the constant folding tried as before but now if the collapse
to constant didn't happen we pass the tree through anyway as the result will be the
same.

In addition, added the missing DB_LTXT_TYPE in the BYTE(ALL) function.

------------------------------------------------------------------------
r3026 | stial01 | 2010-05-19 14:45:21 +0100 (Wed, 19 May 2010) | 21 lines

SIR 121123 Long ID Miscellaneous fixes

1) Core catalog changes to help memory management of long ids:
   Make iiattribute.attname the last column in iiattribute
   Take 4 bytes from relfree, replace with 'relattnametot', which will hold
   size of all column names.
   - change catalog definitions in dmmdata, dmmcre
   - init relattnametot when table is created/altered in dmu routines
   - init relattnametot in upgrade routine for iirelation in dm2d
   - in dm2t.c use relattnametot to allocate space for column names
2) Fixed dbms server exception when multiple dbms servers were started:
   E_RD0125 error (and STACK TRACE from scf). Event data was overflowing stack
   sce_raise when RDF invalidate was raising an event. In sceshare.c increase
   SCEV_BLOCK_MAX to be bigger than than rdf RDR_RB. In sce.c, added a
   consistency check before copying event text.
3) Replaced DB_MAXNAME references not fixed in initial long id code submission
   (dmfjnl.c, dmmalter.c, dm2r.c)
4) Fix stack overflow in crrepkey.sc. MEreq space for DB_MAX_COLS delimited
   column names instead of defining on stack.

------------------------------------------------------------------------