------------------------------------------------------------------------ r3030 | thogo01 | 2010-05-19 22:45:13 +0100 (Wed, 19 May 2010) | 40 lines
The batch processing protocol results in a large number of
messages flowing from the JDBC driver to the Data Access
Server.  The communications protocol permits messages to
be concatenated together in TL level packets.  DAS does
concatenate messages it sends to the driver, but clients
have never taken advantage of this capability for messages
sent to DAS.  There is a long standing bug in DAS which will
result in connection failure if the client concatenates
messages, but which has not been previously discovered.

Batch processing will be more efficient if clients can
utilize the capability to combine the messages in a batch
into a single TL buffer.  There are three parts to this
change:

1.  A one line fix in DAS, read_msg_hdr(), to fix a bug where
    DAS did not properly handle concatenated messages.

2.  Enhance DAS to better handle concatenated messages.
    There are two aspects:

        A.  Rather than separating messages into separate
	    buffers all at once, pull off messages one at
	    a time as needed for processing.  Rather than
	    copying the message to a separate buffer for
	    processing, allocate an RCB with no buffer and
	    reference the message segment where it resides
	    in the original RCB.

        B.  Save and re-use RCB rather than physically
	    freeing.  Two free queues are added to for RCB
	    with no buffers and default sized buffers.

3.  Enhance JDBC driver to concatenate messages in a group
    into a single TL packet.  Due to the bug in DAS, this
    capability is only enabled at the protocol level 7 (and
    higher) where batch processing is supported.

------------------------------------------------------------------------
r3031 | schka01 | 2010-05-20 13:00:12 +0100 (Thu, 20 May 2010) | 3 lines

oj_ijFlagsFile is a hold, not a row;  use the proper index.

------------------------------------------------------------------------