------------------------------------------------------------------------ r3031 | schka01 | 2010-05-20 13:00:12 +0100 (Thu, 20 May 2010) | 3 lines
oj_ijFlagsFile is a hold, not a row;  use the proper index.

------------------------------------------------------------------------
r3032 | schka01 | 2010-05-20 13:00:26 +0100 (Thu, 20 May 2010) | 8 lines

SIR 122513 introduced new data structures that have
to be cloned into a parallel query child for 1:N
parallelism.  Fix opcexch to detect the new qualification
data structures (rows, CX's, and pquals), and fix
qenexch to do the necessary cloning and fixing-up
for the children threads.

------------------------------------------------------------------------