------------------------------------------------------------------------ r3035 | gupsh01 | 2010-05-20 15:15:30 +0100 (Thu, 20 May 2010) | 8 lines
When running sep test dtu54.sep, it was noticed that error
E_AD9998_INTERNAL_ERROR and E_DM93A3_QUALIFY_ROW_ERROR are
reported in the error log. The errors occur because
the minimum length of nchar datatype computed by adi_calclen
was 1. Since length for nchar/nvarchar can only be in
multiples of 2 (sizeof(UCS2)) it is flagged as bad length.

------------------------------------------------------------------------
r3036 | gupsh01 | 2010-05-20 15:30:10 +0100 (Thu, 20 May 2010) | 34 lines

In embedded sql programs, when the buffer size for the
text field of an nvarchar struct declaration is declared
using a #define the length for nvarchar struct passed in
the IIgetdomio() call, for a select into the host
variable, is incorrectly generated by esqlc.

eg:
#define GW155_COL_LEN 5
...
exec sql begin declare section;
nvarchar struct {
    short len;
    wchar_t text[GW155_COL_LEN];
}
exec sql end declare section;

On solaris for the statement

EXEC SQL select ' ' into :res_nvarchar from iidbconstants;

is
...
if (IInextget() != 0) {
    IIgetdomio((short *)0,1,27,
        GW155_COL_LEN + 2, &res_nvarchar);
...
On solaris the expected value is
    IIgetdomio((short *)0,1,27,
        ((GW155_COL_LEN) * 4) + 2, &res_nvarchar);
as sizeof (wchar_t) is 4.

Fixed esqlc to generate the correct length for nvarchar.

------------------------------------------------------------------------