------------------------------------------------------------------------ r3036 | gupsh01 | 2010-05-20 15:30:10 +0100 (Thu, 20 May 2010) | 34 lines
In embedded sql programs, when the buffer size for the
text field of an nvarchar struct declaration is declared
using a #define the length for nvarchar struct passed in
the IIgetdomio() call, for a select into the host
variable, is incorrectly generated by esqlc.

eg:
#define GW155_COL_LEN 5
...
exec sql begin declare section;
nvarchar struct {
    short len;
    wchar_t text[GW155_COL_LEN];
}
exec sql end declare section;

On solaris for the statement

EXEC SQL select ' ' into :res_nvarchar from iidbconstants;

is
...
if (IInextget() != 0) {
    IIgetdomio((short *)0,1,27,
        GW155_COL_LEN + 2, &res_nvarchar);
...
On solaris the expected value is
    IIgetdomio((short *)0,1,27,
        ((GW155_COL_LEN) * 4) + 2, &res_nvarchar);
as sizeof (wchar_t) is 4.

Fixed esqlc to generate the correct length for nvarchar.

------------------------------------------------------------------------
r3037 | thogo01 | 2010-05-20 19:45:14 +0100 (Thu, 20 May 2010) | 26 lines

Batch testing the JDBC driver revealed an edge case where
increasing the batch size by 1 entry per batch at a certain
point resulted in longer execution times rather than the
same or shorter times.  Tracing revealed that the edge case
was the point where the JDBC driver either filled a single
communication buffer or just overflowed a few bytes into a
second buffer.

TCP has an algorithm which delays small packets in situations
where two sends have occured without an intervening receive.
Sending of the short packet is delayed in case there is an
opportunity to combine additional data that might follow with
the current data.

The JDBC driver handles its own buffering and only flushes
buffers when complete or full.  By enabling the TCP NoDelay
option, the delay imposed in the edge case is avoided and
better performance is seen.

In addition to enabling the TCP NoDelay option, the JDBC
driver I/O clases are being enhanced to not flush buffers
when when split.  The flush will only occur when the message
group is complete.

------------------------------------------------------------------------