------------------------------------------------------------------------ r3043 | hanje04 | 2010-05-21 18:30:13 +0100 (Fri, 21 May 2010) | 4 lines
Close hole which can allow symbol.tbl to be updated with "stale" information
resulting in parameters being load.

------------------------------------------------------------------------
r3044 | kiria01 | 2010-05-21 19:02:52 +0100 (Fri, 21 May 2010) | 7 lines

Do not restore .opa_demorgan to TRUE once any negation applied.
The bug only arises if there is a LHS tree with a non-terminal
such as an operator, which then tries to apply the opa_not again.
This might seem more likely to happen but it needs both a boolean
UOP (IS NULL, IS DECIMAL etc) AND the parameter must be an expression
but usually these posfix operators are passed a CONST or a VAR.

------------------------------------------------------------------------