------------------------------------------------------------------------ r3050 | schka01 | 2010-05-25 11:45:16 +0100 (Tue, 25 May 2010) | 20 lines
The current-action tracking in qentproc.c was
inadequate.  Revise so that the current-action
properly tracks through the tproc and properly
handles a reset.  After a RETURN ROW, a tproc
should continue executing with the action
immediately following the RETURN ROW just like
ordinary rowprocs do.

Fix GET so that a GET with no plan (such as a DBP
assignment, or a GET from a SAGG) returns no-rows
after the first execution, until a reset is done.
This fixes issues with tprocs containing nested
FOR loops where an outer loop is driven by a
simple-aggregate query.

Fix goof in the 123565 fix (504676) that broke
tproc queries unless the tproc QP was already in
QSF when the query was issued.

------------------------------------------------------------------------