------------------------------------------------------------------------ r3058 | horda03 | 2010-05-26 12:00:10 +0100 (Wed, 26 May 2010) | 7 lines
Cross integration of change 504908

The IIDBMS.EXE was not being rebuilt when the GCFLIB
was changed. Also the RQFLIB was listed incorrectly.

------------------------------------------------------------------------
r3059 | kiria01 | 2010-05-26 12:54:45 +0100 (Wed, 26 May 2010) | 10 lines

OPF scans for common subexpressions prior to code generation and in this case
spots two cast operators that it deems have the same operants and are thus the
same. Except they are not as the cast operator datatypes have been further resolved
to match the actual length, datatype, precision and collation type of an attribute.

This change makes the commonality test sensitive to the operator return type to
avoid this error and at the same time should reduce some of the tree walking due to
dissimilar subtrees often differing in these result types leading to detecting a
mismatch earlier.

------------------------------------------------------------------------