------------------------------------------------------------------------ r3059 | kiria01 | 2010-05-26 12:54:45 +0100 (Wed, 26 May 2010) | 10 lines
OPF scans for common subexpressions prior to code generation and in this case
spots two cast operators that it deems have the same operants and are thus the
same. Except they are not as the cast operator datatypes have been further resolved
to match the actual length, datatype, precision and collation type of an attribute.

This change makes the commonality test sensitive to the operator return type to
avoid this error and at the same time should reduce some of the tree walking due to
dissimilar subtrees often differing in these result types leading to detecting a
mismatch earlier.

------------------------------------------------------------------------
r3060 | schka01 | 2010-05-26 13:30:15 +0100 (Wed, 26 May 2010) | 13 lines

If DMF attempts to optimize a put coupon
but can't for some legitimate reason (such
as multiple blob columns when only the table
name is supplied), a later double-check was
incorrectly failing.  Make sure that the
double-check isn't tripped.

Also, fix a missing include that was preventing
the blob tests from completing when compiled
32-bit on linux with gcc 4.3.2.  Clean up some
other (symptom-less) missing includes.

------------------------------------------------------------------------