------------------------------------------------------------------------ r3060 | schka01 | 2010-05-26 13:30:15 +0100 (Wed, 26 May 2010) | 13 lines
If DMF attempts to optimize a put coupon
but can't for some legitimate reason (such
as multiple blob columns when only the table
name is supplied), a later double-check was
incorrectly failing.  Make sure that the
double-check isn't tripped.

Also, fix a missing include that was preventing
the blob tests from completing when compiled
32-bit on linux with gcc 4.3.2.  Clean up some
other (symptom-less) missing includes.

------------------------------------------------------------------------