------------------------------------------------------------------------ r3063 | stial01 | 2010-05-26 20:15:11 +0100 (Wed, 26 May 2010) | 6 lines
SIR 121123 Long IDs

Description of changes:
Fix ipm show logging information for databases.

------------------------------------------------------------------------
r3064 | drivi01 | 2010-05-26 21:15:14 +0100 (Wed, 26 May 2010) | 8 lines

Replication doesn't appear to work in Ingres VectorWise installations and is
not supported, remove it from a database tree.

Also, a garbage in the value of the Ingres VectorWise table tree is causing
problems with displaying of tree items.  Stop VDBA from checking the error
when the item is not added, b/c the buffer containing the error is not valid.

------------------------------------------------------------------------