------------------------------------------------------------------------ r3064 | drivi01 | 2010-05-26 21:15:14 +0100 (Wed, 26 May 2010) | 8 lines
Replication doesn't appear to work in Ingres VectorWise installations and is
not supported, remove it from a database tree.

Also, a garbage in the value of the Ingres VectorWise table tree is causing
problems with displaying of tree items.  Stop VDBA from checking the error
when the item is not added, b/c the buffer containing the error is not valid.

------------------------------------------------------------------------
r3065 | stephenb | 2010-05-27 01:20:15 +0100 (Thu, 27 May 2010) | 17 lines

Trac Number: #385

Bug Number: 122320

Release Notes:
rns.start
     (GENERIC)
     Ingres now supports the Daitch-Mokotoff Soundex function through new SQL function SOUNDEX_DM
rns.end

Description of changes:
desc.start
Soundex_dm is added to ADF as a regular Ingres function. The implementation is supported through the
ADF func56ion adu_soundex_dm().
desc.end

------------------------------------------------------------------------