------------------------------------------------------------------------ r3065 | stephenb | 2010-05-27 01:20:15 +0100 (Thu, 27 May 2010) | 17 lines
Trac Number: #385

Bug Number: 122320

Release Notes:
rns.start
     (GENERIC)
     Ingres now supports the Daitch-Mokotoff Soundex function through new SQL function SOUNDEX_DM
rns.end

Description of changes:
desc.start
Soundex_dm is added to ADF as a regular Ingres function. The implementation is supported through the
ADF func56ion adu_soundex_dm().
desc.end

------------------------------------------------------------------------
r3066 | stephenb | 2010-05-27 01:35:16 +0100 (Thu, 27 May 2010) | 21 lines

Bug Number: 123807

Release Notes:
rns.start
     (GENERIC)
     Performance may suffer in some 10.0 applications that are sensitive to the round-trip
     time of GCA communications due to GCA_EOG being set on all GCA messages. This problem
     is now resolved.
rns.end

Description of changes:
desc.start
This change is smarter about when to set GCA_EOG in the DBMS. Rather than setting it for all
non-batched messages, we now set it only for non-batched GCA_RESPONSE and GCA_ERROR messages
(making it clear whether these message types are batched or not). For all other message types
we leave the value unset; this allows GCA to make a decision on EOG based on message type. On
messages that do not terminate the current response-set, EOG will not be set, allowing groups of messages
that belong to the same response-set to be sent in one pass. This returns behavior for non-batched
responses to the same behavior as 9.3 and below.
desc.end
------------------------------------------------------------------------