------------------------------------------------------------------------ r3066 | stephenb | 2010-05-27 01:35:16 +0100 (Thu, 27 May 2010) | 21 lines
Bug Number: 123807

Release Notes:
rns.start
     (GENERIC)
     Performance may suffer in some 10.0 applications that are sensitive to the round-trip
     time of GCA communications due to GCA_EOG being set on all GCA messages. This problem
     is now resolved.
rns.end

Description of changes:
desc.start
This change is smarter about when to set GCA_EOG in the DBMS. Rather than setting it for all
non-batched messages, we now set it only for non-batched GCA_RESPONSE and GCA_ERROR messages
(making it clear whether these message types are batched or not). For all other message types
we leave the value unset; this allows GCA to make a decision on EOG based on message type. On
messages that do not terminate the current response-set, EOG will not be set, allowing groups of messages
that belong to the same response-set to be sent in one pass. This returns behavior for non-batched
responses to the same behavior as 9.3 and below.
desc.end
------------------------------------------------------------------------
r3067 | frima01 | 2010-05-27 09:08:49 +0100 (Thu, 27 May 2010) | 1 line

#607 In fix_dbservice skip update of config file for distributed database, cause there is none. Update unloaddb to create drop link commands. Set tree_convert to true, fixing query trees is neccessary for Ingres 10. Allow only star db owner to upgrade a star db to ensure access to private vnodes.
------------------------------------------------------------------------