------------------------------------------------------------------------ r3069 | schka01 | 2010-05-27 11:45:27 +0100 (Thu, 27 May 2010) | 26 lines
Fix errors in the calls to dmt_show made by
blob processing;  one such error led to the
segv seen in SD issue 144789.
Fix minor errors in blob-put optimization
detection in the sequencer, it needs to pass
blank-padded table name/owner if it's going to
pass anything.

Having gotten blob put optimization to actually
kick in (as verified in the debugger), turn it
off.  :-(
Put optimization cannot be used at present because
the optimizer can compile in LVCH-to-LVCH coercions,
and put optimization is based on the assumption
that the blob coupon will be sent straight to DMF.
The cure is probably to detect a no-op blob
move in DMF and simply transfer the coupon,
but detecting that reliably is a difficult
problem.

Please note that COPY FROM put optimization remains
in place, as the optimizer is not involved with
COPY FROM.  QEF guarantees that the coupons go
directly to DMF.

------------------------------------------------------------------------
r3070 | kiria01 | 2010-05-27 17:15:12 +0100 (Thu, 27 May 2010) | 8 lines

QA - integer casts with NVARCHAR(1) return 0 for all values

The sense was interchanged for the temporary length calculation for
NCHAR and NVARCHAR resulting in NVARCHAR(1) trying to use a VARCHAR(0)
temporary - hence the 0 result.

------------------------------------------------------------------------