------------------------------------------------------------------------ r3077 | hanje04 | 2010-05-28 21:45:15 +0100 (Fri, 28 May 2010) | 17 lines
Fix up GUI installer for IVW after first round of QA:

 - Mantis 730 - Add on-line help for new IVW config
 - Mantis 731 - Make sure right info is displayed in pre/post install summary
 - Mantis 732 - Remove ABF from default package list for IVW
 - Mantis 708 - Replace Vectorwise with VectorWise in GUI

Other changes requested/required:

 - Make sure Next button is correctly disabled for the license screen
   in "New Install" mode.
 - Bump numeric limits for memory and disk size parameters to 16384
 - Set group_size minimum to 1
 - Round up BlockSize and GroupSize to a power of 2

------------------------------------------------------------------------
r3078 | lunbr01 | 2010-05-28 23:45:14 +0100 (Fri, 28 May 2010) | 17 lines

This change fixes a minor problem introduced by enhancement 122679.
A check had been added to listen completion to detect a mismatch
between the owner of the incoming pipe (according to the OS) and
the user id sent by the client in the connection peer info message.
The only action taken at this level, if a mismatch was detected, was
to issue a warning trace message (if GC tracing on) to highlight a
possible security anomaly (GCS does similar checking and will
generate an error if appropriate).  The warning trace message is
erroroneously being issued when Ingres is run as a service and
the pipe owner returned by the OS is in uppercase.  The userid
sent in the GC peer info message by the Windows named-pipe client
process is always in lowercase.  Since userids on Windows are
case-insensitive, and in order to be consistent with other places
in the GCA CL (such as GCusername()), the fix is to lowercase the
pipe owner, thus eliminating the bogus mismatch.

------------------------------------------------------------------------