------------------------------------------------------------------------ r3078 | lunbr01 | 2010-05-28 23:45:14 +0100 (Fri, 28 May 2010) | 17 lines
This change fixes a minor problem introduced by enhancement 122679.
A check had been added to listen completion to detect a mismatch
between the owner of the incoming pipe (according to the OS) and
the user id sent by the client in the connection peer info message.
The only action taken at this level, if a mismatch was detected, was
to issue a warning trace message (if GC tracing on) to highlight a
possible security anomaly (GCS does similar checking and will
generate an error if appropriate).  The warning trace message is
erroroneously being issued when Ingres is run as a service and
the pipe owner returned by the OS is in uppercase.  The userid
sent in the GC peer info message by the Windows named-pipe client
process is always in lowercase.  Since userids on Windows are
case-insensitive, and in order to be consistent with other places
in the GCA CL (such as GCusername()), the fix is to lowercase the
pipe owner, thus eliminating the bogus mismatch.

------------------------------------------------------------------------
r3079 | stephenb | 2010-05-31 01:24:03 +0100 (Mon, 31 May 2010) | 24 lines

Bug Number: 123826

Release Notes:
rns.start
     (GENERIC)
     An incorrectly coded Java program that repeatedly uses addBatch to execute
     a prepared insert statement without re populating the query parameters may
     cause a SEGV or Access Violation in the DBMS.
rns.end

Description of changes:
desc.start
When the parameters are not correctly re-populated in the Java program it causes the
application to send the query with an incorrect number of parameters. The copy
optimization for batched prepared queries bypasses the parser for this special
case, but it was not performing a sanity check on the query parameters before trying
to execute the query (this would have ordinarily been done in the parser). As a result
it tried to execute the query with an incorrect parameter list which caused a SEGV in the
QEF copy code. This change adds a sanity check which now generates a user error and
aborts the query.

This change resolves issue 144218
desc.end
------------------------------------------------------------------------