------------------------------------------------------------------------ r3090 | hanal04 | 2010-06-02 09:15:22 +0100 (Wed, 02 Jun 2010) | 9 lines
Cross integration of change 505041

ima_start_sampler() and ima_stop_sampler() both SEGV in CVlower()
because they pass a string constant into CSmonitor(). This
change moves the command strings into a local buffer so that
CVlower() can be safely called.

------------------------------------------------------------------------
r3091 | hanal04 | 2010-06-02 09:30:28 +0100 (Wed, 02 Jun 2010) | 19 lines

Cross integration of change 505043

In opc_fatts() we choose a row for the fatts to go into (indexed
by qp_row_cnt). In order to allocate the row (and bump the
qp_row_cnt) we need to call opc_ptrow(). However, opc_fatts()
was selecting the row and then calling opc_adend() before
calling opc_ptrow().
opc_adend() may also choose and allocate rows in the DSH by
calling opc_ptrow(). For this bug opc_adend() was allocating a
row for the repeat query parameter. By calling opc_adend() before
opc_ptrow() (for the fatts row) we were allocating the same
row for the QEN_PARM as selected for the fatts.
This change moves the call to opc_adend() below the call to
opc_ptrow() in opc_fatts() to make sure the row for the fatts
is allocated and the qp_row_cnt index bumped before any
further allocations take place.

------------------------------------------------------------------------