------------------------------------------------------------------------ r3099 | jenjo02 | 2010-06-03 18:00:32 +0100 (Thu, 03 Jun 2010) | 10 lines
If a statement causes an "internal" (QEF_ITRAN) transaction to be started in a MVCC-likely
database, a CRIB is not allocated.

If the transaction is later uprated to QEF_MSTRAN and used for MVCC, a SEGV
will be thrown in dmt_open() due to the lack of a CRIB.

This change to dmx_begin() ignores the setting of or lack of DMX_USER_TRAN
when deciding whether or not to allocate a CRIB.

------------------------------------------------------------------------
r3100 | hanje04 | 2010-06-03 19:15:13 +0100 (Thu, 03 Jun 2010) | 3 lines

Correct typo in ingres.template which always forces upgrade mode.

------------------------------------------------------------------------